+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * implement simple regex based syntax-extension
 * performance improvements
 * new interpreter with TCO and Continuations inspired by js-scheme [#127](https://github.com/jcubic/lips/issues/127)
+* new higher order function `matcher` that return function that check if object is the same
 ### Bugfix
 * fix doc string for `make-rectangular`
 * `-inf.0`/`+inf.0` are now real lips numbers